update version 1 to support the following:  1- support more charset 0x33D, 0x33E, 0x33F, 0x340, 0x352, 0x353, 0x354  2- add prefetch rows url option  3- add IStringConverter for custom encode and decode  4- fix rowid reading

+++ network/redirect_packet.go
@@ -31,7 +31,7 @@ func newRedirectPacketFromData(packetData []byte) *RedirectPacket {
 	pck := RedirectPacket{
 		packet: Packet{
 			dataOffset: 10,
-			length:     uint32(binary.BigEndian.Uint16(packetData)),
+			length:     binary.BigEndian.Uint16(packetData),
 			packetType: PacketType(packetData[4]),
 			flag:       packetData[5],
 		},
